feat(douyinopen): 升级公共组件

+++ src/SKIT.FlurlHttpClient.ByteDance.DouyinOpen/Models/DataExternal/Billboard/HotVideo/DataExternalBillboardHotVideoResponse.cs
@@ -81,7 +81,7 @@
                         /// </summary>
                         [Newtonsoft.Json.JsonProperty("hot_value")]
                         [System.Text.Json.Serialization.JsonPropertyName("hot_value")]
-                        public double HotValue { get; set; }
+                        public decimal HotValue { get; set; }
                     }
                 }
