+++ src/SKIT.FlurlHttpClient.ByteDance.DouyinOpen/Models/GoodLife/Trade/GoodLifeTradeOrderQueryV1Response.cs
@@ -229,14 +229,14 @@ namespace SKIT.FlurlHttpClient.ByteDance.DouyinOpen.Models
                                 /// </summary>
                                 [Newtonsoft.Json.JsonProperty("lat")]
                                 [System.Text.Json.Serialization.JsonPropertyName("lat")]
-                                public double? Latitude { get; set; }
+                                public decimal? Latitude { get; set; }
 
                                 /// <summary>
                                 /// 获取或设置经度。
                                 /// </summary>
                                 [Newtonsoft.Json.JsonProperty("lng")]
                                 [System.Text.Json.Serialization.JsonPropertyName("lng")]
-                                public double? Longitude { get; set; }
+                                public decimal? Longitude { get; set; }
 
                                 /// <summary>
                                 /// 获取或设置脱敏后的收货人姓名。
